backend/requirements.txt
# FastAPI and Web Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database - using older psycopg2-binary version with prebuilt wheels
sqlalchemy==2.0.27
psycopg2-binary>=2.9.9
alembic==1.13.1

# OCR and Image Processing
paddleocr==3.0.0
paddlepaddle==3.2.0
Pillow>=11.0.0,<12
pdf2image==1.17.0

# Data Validation and Serialization
pydantic==2.6.1
pydantic-settings==2.1.0

# Encryption and Security
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1

# Excel Processing
openpyxl==3.1.2
xlsxwriter==3.2.0

# Utilities
python-dateutil==2.8.2
regex==2023.12.25
